create features and shared dir, move todolist components into new dir, update imports
