+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css';
-import TodoList from './TodoList';
+import TodoList from './features/TodoList/TodoList';
 import TodoForm from './TodoForm';
 import { useState } from 'react';
